+ highcharts - bizcharts. refactored react frontend. made express requests => graph
