+++ .gitignore
@@ -7,7 +7,7 @@ logs
 
 # dependencies
 /node_modules
-
+/client/node_modules
 # testing
 /coverage
