+++ README.md
@@ -20,37 +20,27 @@ Sample Response:
     }
 }
 ```
+# run 
+
+* nodemon server.js
+* cd client
+* npm start
+
 
 # important files
 
-* pullData.js: hits the Infura API for all the transactions in the latest block and saves them to a local postgres DB. 
+* pullData.js: called by cronjob. polls infura api for new blocks to save in DB
 * frequency.js: returns contract tx frequency by timestamp
 * server.js: entry point 
+* home.js: chart controller
 
-# Server
-
-```
-npm i
-sudo -u postgres createdb tx
-psql -U postgres -d tx -f sql/makeTable.sql
-PGUSER='postgres' PGDATABASE='tx' PGPORT=5432 node pullData.js p 1
-PGUSER='postgres' PGDATABASE='tx' PGPORT=5432 node server.js
-```
-
-# Client
-```
-npm run client
-```
 
 # next
 
-- Build Frontend with BizCharts. coinmarketcap.com for reference. @eswarasai helping here  
-
-# future ideas
-
-- scrape dapp contract names from etherscan?
+- input form for contractID and dates. multiple graphs. include functions in graph
+- gather more contractID <> dApp mappings
 - handle failed transactions. some blocks have tons https://etherscan.io/txs?block=4905182&p=1.
     to do this, explore transaction receipts trie https://github.com/ethereum/wiki/wiki/JSON-RPC#eth_getblockbyhash
-- decentralized web stack? ipfs? 
+- put on ipfs
 
 inspirated by @owocki https://github.com/gitcoinco/skunkworks/issues/19#issuecomment-354695051
