+++ client/src/containers/index.js
@@ -1,0 +1,1 @@
+export { default as HomeContainer } from './Home';
