+++ client/src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import { routerReducer } from 'react-router-redux';
+
+export default combineReducers({
+  routing: routerReducer
+});
