+++ client/src/routes.js
@@ -1,0 +1,10 @@
+import * as Containers from './containers';
+
+
+export const routes = [
+  {
+    component: Containers.HomeContainer,
+    exact: true,
+    path: '/',
+  }
+];
