+++ client/src/store.js
@@ -1,0 +1,19 @@
+import { createStore, applyMiddleware, compose } from 'redux';
+import thunk from 'redux-thunk';
+import { routerMiddleware } from 'react-router-redux';
+import { createLogger } from 'redux-logger';
+import createHistory from 'history/createBrowserHistory';
+
+import reducers from './reducers';
+
+export const history = createHistory();
+
+// Redux DevTools
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+
+const store = createStore(
+  reducers,
+  composeEnhancers(applyMiddleware(thunk, createLogger(), routerMiddleware(history)))
+);
+
+export default store;
