+++ package.json
@@ -8,10 +8,13 @@
     "abi-decoder": "^1.0.9",
     "antd": "^3.1.6",
     "async-request": "^1.2.0",
+    "axios": "^0.18.0",
     "bizcharts": "^3.1.3-beta.2",
     "bizcharts-plugin-slider": "^2.0.1",
     "dotenv": "^1.0.0",
     "express": "^4.16.2",
+    "highcharts": "^6.0.7",
+    "highcharts-react-official": "github:highcharts/highcharts-react",
     "history": "^4.7.2",
     "moment": "^2.20.1",
     "node-pre-gyp": "^0.6.39",
@@ -19,6 +22,7 @@
     "prettier": "^1.10.2",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
+    "react-highcharts": "^16.0.2",
     "react-redux": "^5.0.2",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
@@ -44,7 +48,7 @@
     "eslint-plugin-react": "^7.5.1"
   },
   "lint-staged": {
-    "src/**/*.{js,jsx,json,css}": [
+    "**/*.{js,jsx,json,css}": [
       "prettier --single-quote --write",
       "git add"
     ]
@@ -55,5 +59,6 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
-  }
+  }, 
+  "proxy": "http://localhost:5000/"
 }
