--- src/containers/Home.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-
-import SliderChart from '../components/SliderChart';
-
-
-class Home extends Component {
-  render() {
-    return (
-      <SliderChart/>
-    );
-  }
-}
-
-
-export default Home;
