--- src/containers/index.js
@@ -1,1 +1,0 @@
-export { default as HomeContainer } from './Home';
