--- src/reducers/index.js
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import { routerReducer } from 'react-router-redux';
-
-export default combineReducers({
-  routing: routerReducer
-});
