--- src/routes.js
@@ -1,10 +1,0 @@
-import * as Containers from './containers';
-
-
-export const routes = [
-  {
-    component: Containers.HomeContainer,
-    exact: true,
-    path: '/',
-  }
-];
